+++ packages/frontend/src/pages/scaling/liveness/index.tsx
@@ -1,16 +1,13 @@
-import { LivenessApiResponse } from '@l2beat/shared-pure'
 import React from 'react'
 
 import { Config } from '../../../build/config'
 import { PageWrapper } from '../../../components'
+import { PagesData } from '../../Page'
 import { getProps } from './props/getProps'
 import { ScalingLivenessPage } from './view/ScalingLivenessPage'
 
-export function getLivenessPage(
-  config: Config,
-  livenessData: LivenessApiResponse,
-) {
-  const { props, wrapper } = getProps(config, livenessData)
+export function getLivenessPage(config: Config, pagesData: PagesData) {
+  const { props, wrapper } = getProps(config, pagesData)
   return {
     slug: '/scaling/liveness',
     page: (
